update version code
