+++ Cargo.toml
@@ -1,12 +1,12 @@
 [workspace]
 members = ["tower-mqtt","tower-raft","tower"]
 
-
 [workspace.package]
 name = "ThingsTower"
 version = "0.1.0"
 edition = "2021"
 authors = ["gsxhnd"]
+homepage = ""
 description = "MQTT Server for v3.1, v3.1.1 and v5.0 protocols"
 repository = "https://github.com/gsxhnd/ThingsTower"
 license = "MIT license"
@@ -17,13 +17,13 @@ rust-version = "1.68"
 
 
 [path]
-tower_mqtt = {path ="tower-mqtt"}
-tower_raft = {path ="tower-raft"}
+tower_mqtt = { path ="tower-mqtt" }
+tower_raft = { path ="tower-raft" }
 
 [workspace.dependencies]
 axum = {version = "0.6"}
 bytes = { version = "1.4", features = ["serde"] }
-clap = { version = "4.2.7", features = ["derive","env"] }
+clap = { version = "4.3", features = ["derive","env"] }
 futures = { version = "0.3" }
 futures-util = { version = "0.3" }
 mqttrs = { version = "0.4.1", features = [ "derive" ] }
