+++ tower-mqtt/src/lib.rs
@@ -2,12 +2,11 @@ pub mod v3;
 pub mod v5;
 
 mod config;
-mod decoder;
 mod server;
 mod topic;
 mod types;
-mod version;
+
+pub(self) mod version;
 
 pub use self::config::MqttConfig;
 pub use self::server::MqttServer;
-// pub use self::types::QoS;
